Made functions a lot neater, and added MANY comments, still need to bug fix some thigns
